Split birthday into two files
